+++ Pandora/Paradigm/Basis.hs
@@ -1,6 +1,6 @@
 module Pandora.Paradigm.Basis (module Exports, note, hush, left, right, this, that) where
 
-import Pandora.Paradigm.Basis.Twister as Exports
+import Pandora.Paradigm.Basis.Construction as Exports
 import Pandora.Paradigm.Basis.Instruction as Exports
 import Pandora.Paradigm.Basis.Fix as Exports
 import Pandora.Paradigm.Basis.Yoneda as Exports